fix : StudyGoal, AttendanceCheck 클래스 수정 및 리팩토링

+++ src/main/java/com/planner/server/domain/attendance_check/service/AttendanceService.java
@@ -1,8 +1,7 @@
 package com.planner.server.domain.attendance_check.service;
 
 import com.planner.server.domain.attendance_check.dto.AttendanceDto;
-import com.planner.server.domain.attendance_check.dto.AttendanceResDtoList;
-import com.planner.server.domain.attendance_check.dto.AttendanceResDto;
+import com.planner.server.domain.attendance_check.dto.AttendanceListDto;
 import com.planner.server.domain.attendance_check.entity.AttendanceCheck;
 import com.planner.server.domain.attendance_check.repository.AttendanceRepository;
 import com.planner.server.domain.user.dto.UserDto;
@@ -12,8 +11,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 @Service
 @RequiredArgsConstructor
@@ -21,13 +19,15 @@ public class AttendanceService {
 
     private final AttendanceRepository attendanceRepository;
     private final UserRepository userRepository;
-    public AttendanceDto saveByUserId(String userId) {
-        User user = userRepository.findById(UUID.fromString(userId)).get();
-
-        UUID id = UUID.randomUUID();
+    public AttendanceDto save(UUID userId) {
+        Optional<User> findUser = userRepository.findById(userId);
+        if(!findUser.isPresent()){
+            throw new NoSuchElementException("[id] 확인 요망.");
+        }
+        User user = findUser.get();
 
         AttendanceCheck attendanceCheck = AttendanceCheck.builder()
-                .id(id)
+                .id(UUID.randomUUID())
                 .user(user)
                 .createdAt(LocalDateTime.now())
                 .build();
@@ -35,15 +35,25 @@ public class AttendanceService {
         return AttendanceDto.toDto(saveAttendance, UserDto.toDto(user));
     }
 
-    public AttendanceResDtoList findByUserId(String userId) {
-        List<AttendanceCheck> attendanceList = attendanceRepository.findByUserId(UUID.fromString(userId));
-        AttendanceResDtoList attendanceResDtoList = AttendanceResDtoList.toDtoList(attendanceList);
+    public AttendanceListDto findByUserId(UUID userId) {
+        List<AttendanceCheck> attendanceList;
+        try{
+            attendanceList = attendanceRepository.findByUserId(userId);
+        }
+        catch (NoSuchElementException e){
+            throw new NoSuchElementException("[id]값 확인 요망.");
+        }
+        AttendanceListDto attendanceResDtoList = AttendanceListDto.toDtoList(attendanceList);
         return attendanceResDtoList;
     }
 
     // 쿼리가 2번 생성됨. (수정 필요)
-    public void deleteById(String id) {
-        AttendanceCheck attendanceCheck = attendanceRepository.findById(UUID.fromString(id)).get();
+    public void deleteById(UUID id) {
+        Optional<AttendanceCheck> byId = attendanceRepository.findById(id);
+        if(!byId.isPresent()){
+            throw new NoSuchElementException("[id]값 확인 요망");
+        }
+        AttendanceCheck attendanceCheck = byId.get();
         attendanceRepository.delete(attendanceCheck);
     }
 }
